requirements-dev.txt
pydantic
requests
uuid
gTTS
zhon
jieba
azure-cognitiveservices-speech
pymongo
python-dotenv
azure-identity
azure-storage-blob
matplotlib
pytest
black
streamlit
streamlit-elements
streamlit-mic-recorder
werkzeug
faker
bs4
cryptography
google-cloud-translate
google-cloud-aiplatform
google-cloud-texttospeech
google-cloud-storage
google-generativeai
google-cloud-firestore
jupyter
ipykernel
ipywidgets